update basemodel and remove test

+++ basemodel_class/basemodel_collection.py
@@ -4,4 +4,4 @@ from pydantic import BaseModel
 class Data(BaseModel):
     temperature: int
     humidity: int
-    co: float
+    co: int
